+++ src/main/java/com/projectlyrics/server/domain/user/usecase/command/UserAuthUseCase.java
@@ -1,9 +1,9 @@
-package com.projectlyrics.server.domain.user.service;
+package com.projectlyrics.server.domain.user.usecase.command;
 
 import com.projectlyrics.server.domain.user.dto.response.LoginResponse;
 import com.projectlyrics.server.domain.auth.external.dto.request.UserLoginRequest;
 
-public interface UserCommandService {
+public interface UserAuthUseCase {
 
   LoginResponse login(UserLoginRequest request);
 }
